Added 'fastMode' to aggressively empty the notification buffer
Enabling this option may not be advisable if node-apn runs within a larger application in-case it stalls while emptying a large buffer. In the case of high throughput it would be sensible to run node-apn within a separate worker process servicing a queue in which case the `fastMode` should be enabled.